Fix 'SYCL feature test compile failed' problem
Add _get_xfail_test() to _helper.py

_is_type() --> _match_type()

Fix docstring

Set CMAKE_C##_COMPILER in setup.py

Revert setup.py
